@@ -4,9 +4,9 @@
     "type": "library",
     "license": "LGPL-3.0",
     "require": {
-        "php": "^8.0",
-        "phpstan/phpstan": "^1.8.3",
-        "nikic/php-parser": "^4.15.1"
+        "php": "^8.1",
+        "phpstan/phpstan": "1.10.*",
+        "nikic/php-parser": "4.17.*"
     },
     "autoload": {
         "psr-4": {
